[feat] 댓글 삭제 시 권한 확인 추가

+++ src/main/java/com/example/smart_edu_team/comment/CommentController.java
@@ -56,6 +56,9 @@ public class CommentController {
      */
     @PostMapping("/{postId}/delete/{id}")
     public String deleteComment(@PathVariable Long postId, @PathVariable Long id, Principal principal) {
+        if (!Objects.equals(principal.getName(), principal.getName()) && !Objects.equals(principal.getName(), "admin")) {
+            return "post/bad_request";
+        }
         commentService.deleteComment(id);
         return "redirect:/posts/detail/" + postId.toString();
     }
